Fix error for style.css and ng test
This commit fixes the 'Could not resolve' error for the styles.css file that appeared after merging the main and master branches. The error was fixed after reinstalling Angular Material. Moreover, it fixes the error after running the
g test command by updating the include array in the 	sconfig.spec.json file.